% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_cost_by_city.R
\name{avg_cost_by_city}
\alias{avg_cost_by_city}
\title{Average trip cost for each city}
\usage{
avg_cost_by_city(data)
}
\arguments{
\item{data}{A data frame that has columns called city and total_cost.}
}
\value{
A data frame with two columns: city and average_cost.
}
\description{
This function groups trips by city and finds the average total_cost
for each city in the data.
}
\examples{
data(travel_data)
avg_cost_by_city(travel_data)

}
