% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cost_by_city.R
\name{plot_cost_by_city}
\alias{plot_cost_by_city}
\title{Plot average trip cost by city}
\usage{
plot_cost_by_city(data)
}
\arguments{
\item{data}{A data frame that has columns called city and total_cost.}
}
\value{
A ggplot object showing a bar chart of average cost by city.
}
\description{
This function makes a bar chart that shows the average trip
cost for each city in the data.
}
\examples{
data(travel_data)
plot_cost_by_city(travel_data)

}
