% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_summary.R
\name{print.trip_summary}
\alias{print.trip_summary}
\title{Print method for trip_summary objects}
\usage{
\method{print}{trip_summary}(x, ...)
}
\arguments{
\item{x}{A trip_summary object.}

\item{...}{Extra arguments (not used).}
}
\value{
The same object, invisibly.
}
\description{
This function controls how trip_summary objects
are printed to the screen.
}
\examples{
data(travel_data)
ts <- trip_summary(travel_data)
print(ts)

}
