% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_duration.R
\name{trip_duration}
\alias{trip_duration}
\title{Calculate trip length in days}
\usage{
trip_duration(data)
}
\arguments{
\item{data}{A data frame that has columns called start_date and end_date.}
}
\value{
A numeric vector with the length of each trip in days.
}
\description{
This function uses the start_date and end_date columns
to find how many days each trip lasted.
}
\examples{
data(travel_data)
trip_duration(travel_data)

}
