% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_summary.R
\name{trip_summary}
\alias{trip_summary}
\title{Create a simple trip summary object}
\usage{
trip_summary(data)
}
\arguments{
\item{data}{A data frame like travel_data.}
}
\value{
An object of class "trip_summary".
}
\description{
This function makes a special object that stores
basic summaries of the travel data. It uses the
functions from this package.
}
\examples{
data(travel_data)
ts <- trip_summary(travel_data)
ts

}
